requirements.txt
keras==2.8.0
numpy==1.22.3
pandas==1.4.2
PyYAML==6.0
tensorflow==2.8.0
